project_root/requirements.txt
tensorflow>=2.15.0
torch>=2.2.1
scikit-learn>=1.4.0
nltk>=3.8.1
spacy>=3.7.4
opencv-python>=4.9.0.80
moviepy>=1.0.3
flask>=3.0.2
fastapi>=0.110.0
lime>=0.2.0.1
shap>=0.44.0
tweepy>=4.14.0
optuna>=3.6.0
pytest>=8.0.2
requests>=2.31.0
transformers>=4.39.3
facebook-sdk>=3.1.0